db_handling

Module bitcoin

Bitcoin key derivation using BIP-32/44/84/86

This implements key derivation for Bitcoin with support for:

BIP-44: Legacy addresses (P2PKH) - m/44’/0’/account’/change/index

BIP-84: Native SegWit addresses (P2WPKH) - m/84’/0’/account’/change/index

BIP-86: Taproot addresses (P2TR) - m/86’/0’/account’/change/index

Mainnet uses coin type 0, testnet uses coin type 1.

Structs

BitcoinKeyPair

Bitcoin key pair with secp256k1 keys

PsbtInfo

Summary of a PSBT for user review

PsbtInputInfo

Information about a PSBT input for display

PsbtOutputInfo

Information about a PSBT output for display

Constants

COIN_TYPE_MAINNET

SLIP-0044 coin types

COIN_TYPE_TESTNET

PURPOSE_LEGACY

BIP-44 purpose constants

PURPOSE_SEGWIT

PURPOSE_TAPROOT

Functions

analyze_psbt

Analyze a PSBT and extract display information

derive_bitcoin_key

Derive Bitcoin key from seed phrase

derive_segwit_key

Derive a Native SegWit (BIP-84) key Path: m/84’/0’/account’/change/index

derive_taproot_key

Derive a Taproot (BIP-86) key Path: m/86’/0’/account’/change/index

export_xpub

Export extended public key (xpub/ypub/zpub) for watch-only wallet

finalize_psbt

Finalize a fully-signed PSBT and extract the transaction

format_btc

Format satoshis as BTC string

format_sats

Format satoshis for display

get_bitcoin_address

Retrieve a Bitcoin address for a given public key hex

parse_bitcoin_path

Parse Bitcoin derivation path

parse_psbt_base64

Parse a PSBT from base64 string

parse_psbt_bytes

Parse a PSBT from raw bytes

psbt_to_base64

Serialize a PSBT to base64 string

psbt_to_hex

Serialize a PSBT to hex string

sign_psbt

Sign a PSBT with the given key pair

store_bitcoin_address

Store a Bitcoin address for a given public key hex