db_handling

Module cold_default

Making and restoring cold database with default content

Cold database is the database that is operated from inside the Vault.

Cold database contains following trees:

ADDRTREE with public addresses data

HISTORY with Vault history log

METATREE with network metadata

SETTREE with settings: types information, Vault dangerous exposures record and Vault database general verifier

SPECSTREE with network specs

TRANSACTION for temporary storage of the transaction data

VERIFIERS with network verifiers data

For release, the cold database is generated on the hot side and then copied verbatim into Vault files during the build.

Before the database could be used, it must be initiated:

History log old entries (if any are present) are removed and a new entry Event::DatabaseInitiated is added

General verifier is set and this event is recorded in the history log. By default, Vault sets up Parity-associated key as a general verifier. This could be later on changed by the user.

Vault then reads and updates the database as it operates.

There are two ways to reset the database from the Vault user interface. Either would remove all the keys and restore the database to the release state. The difference would be only in the general verifier setting:

Wipe all data would set the general verifier to default Some(_), with Parity-associated key inside

Remove general certificate would set the general verifier to None. User would then be able to set up own general verifier, preferably immediately afterwards, by loading to Vault any verified data. Setting up a new general verifier would remove all data associated with the general verifier from the Vault database to avoid confusion as to who verified what information.

Functions

init_db

Initiate cold database and set up the database general verifier to given [Verifier].

populate_all_network_specs

Generate database with networks specs both for substrate and ethereum

populate_cold

Generate initiated test cold database with default content, and create in it Alice default addresses.

populate_cold_nav_test

Generate not initiated test cold database for navigator testing.

populate_cold_nav_test_with_ethereum_based_networks

Generate not initiated test cold database for navigator testing with Mythos network.

populate_cold_no_metadata

Generate initiated test cold database without network metadata.

populate_cold_no_networks

Generate initiated test cold database with no network-associated data.

signer_init_no_cert

Initiate Vault database with general verifier set up to Verifier(None).

signer_init_with_cert

Initiate Vault database with default general verifier (Parity-associated key).