Struct db_handling::db_transactions::TrDbColdStub

pub struct TrDbColdStub { /* private fields */ }

Draft for cold database atomic transaction, constructed for Vault update transaction (add_specs, load_metadata, load_types).

TrDbColdStub is stored SCALE-encoded in TRANSACTION tree of the cold database under key STUB while the update is considered by the user. Draft is applied atomically to the cold database if the update is accepted.

Accepting an update could result in adding or removing database data.

TrDbColdStub contains [Event] set for HISTORY tree update and BatchStub update drafts with corresponding removals and additions for database trees:

ADDRTREE

METATREE

SPECSTREE

SETTREE

VERIFIERS

Note that all the checking is done before the TrDbColdStub is written into TRANSACTION tree, apply method will check only that the checksum known to the user is the same as the one database has currently.

Implementations

impl TrDbColdStub

pub fn new() -> Self

Construct new empty TrDbColdStub.

pub fn from_storage(database: &Db, checksum: u32) -> Result<Self>

Recover TrDbColdStub from storage in the cold database.

Function requires correct checksum to make sure the transaction is still the one that was shown to the user previously, and no changes to the database have occured after the atomic transaction draft was placed into storage.

TRANSACTION tree is cleared in the process.

pub fn store_and_get_checksum(&self, database: &Db) -> Result<u32>

Put SCALE-encoded TrDbColdStub into storage in the TRANSACTION tree of the cold database under the key STUB.

Function returns u32 checksum. This checksum is needed to recover stored TrDbColdStub using from_storage method.

The TRANSACTION tree is cleared prior to adding data to storage.

pub fn new_history_entry(self, event: Event) -> Self

Add new [Event] in history_stub field of the TrDbColdStub

pub fn add_metadata(self, meta_values: &MetaValues) -> Self

Prepare adding the metadata received as load_metadata update into the cold database:

Add a (key, value) pair to the metadata additions queue in metadata_stub. Key is [MetaKey] in key form, value is metadata in Vec<u8> format.

Add corresponding Event::MetadataAdded(_) into history_stub.

pub fn remove_metadata(self, meta_values: &MetaValues) -> Self

Prepare removing the metadata from the cold database:

Add [MetaKey] in key form to the metadata removals queue in metadata_stub.

Add corresponding Event::MetadataRemoved(_) into history_stub.

Function is used for Hold and GeneralHold processing when, respectively, the network verifier or the general verifier is changed.

pub fn add_network_specs(
self,
database: &Db,
network_specs_to_send: &NetworkSpecs,
valid_current_verifier: &ValidCurrentVerifier,
general_verifier: &Verifier
) -> Result<Self>

Prepare adding [OrderedNetworkSpecs] into the cold database:

Transform received in add_specs payload [NetworkSpecs] into [OrderedNetworkSpecs] by adding order field. Networks are always added in the end of the network list, with order set to the total number of network specs entries currently in Vault. When a network is removed, the order of the remaining networks gets rearranged, see details in function remove_network.

Add a (key, value) pair to the network specs additions queue in network_specs_stub. Key is [NetworkSpecsKey] in key form, value is SCALE-encoded [OrderedNetworkSpecs].

Add corresponding Event::NetworkSpecsAdded(_) into history_stub.

Add root address for the network if the AddressDetails entry with matching Encryption already exists, i.e. add (key, value) pair to the address additions queue in addresses_stub. Key is AddressKey in key form, value is SCALE-encoded updated AddressDetails.

If address was added, add corresponding Event::IdentityAdded(_) into history_stub.

Note that add_network_specs does not deal with network verifiers: verifier data is not necessarily updated each time the network specs are added.

pub fn remove_network_specs(
self,
network_specs: &OrderedNetworkSpecs,
valid_current_verifier: &ValidCurrentVerifier,
general_verifier: &Verifier
) -> Self

Prepare removing [OrderedNetworkSpecs] from the cold database:

Add [NetworkSpecsKey] in key form to the network specs removal queue in network_specs_stub.

Add corresponding Event::NetworkSpecsRemoved(_) into history_stub.

Function is used for Hold and GeneralHold processing when, respectively, the network verifier or the general verifier is changed.

Note that function does not deal with the verifiers nor with the addresses.

Verifiers remain unchanged during the hold processing.

The addresses are not removed and will be again visible from the user interface when the properly verified network specs are loaded in Vault.

pub fn new_general_verifier(self, general_verifier: &Verifier) -> Self

Prepare adding new general verifier [Verifier] into the cold database:

Add a (key, value) pair to the settings additions queue in settings_stub. Key is GENERALVERIFIER and the value is SCALE-encoded [Verifier] that is set to be the new general verifier.

Add corresponding Event::GeneralVerifierSet(_) into history_stub.

pub fn add_types(
self,
types: &ContentLoadTypes,
general_verifier: &Verifier
) -> Self

Prepare adding types information [ContentLoadTypes] received as load_types update into the cold database:

Add a (key, value) pair to the settings additions queue in settings_stub. Key is TYPES and the value is [ContentLoadTypes] types information in store format (SCALE-encoded).

Add corresponding Event::TypesAdded(_) into history_stub.

pub fn remove_types(
self,
types: &ContentLoadTypes,
general_verifier: &Verifier
) -> Self

Prepare removing types information from the cold database:

Add TYPES key to the settings removal queue in settings_stub.

Add corresponding Event::TypesRemoved(_) into history_stub.

Function is used to process GeneralHold when general verifier is changed.

pub fn new_network_verifier(
self,
verifier_key: &VerifierKey,
valid_current_verifier: &ValidCurrentVerifier,
general_verifier: &Verifier
) -> Self

Prepare adding new network verifier [ValidCurrentVerifier] into the cold database:

Add a (key, value) pair to the verifiers additions queue in verifiers_stub. Key is [VerifierKey] and the value is SCALE-encoded [ValidCurrentVerifier] that is set to be the new verifier for the network.

Add corresponding Event::NetworkVerifierSet(_) into history_stub.

pub fn apply(self, database: &Db) -> Result<()>

Transform TrDbColdStub into TrDbCold and apply to the database with a given name, in a single transaction.

The TRANSACTION tree gets cleared in the process.

It is unlikely that this clearing is ever doing anything, as the intended use of the TrDbColdStub is to recover it from the database (with clearing the TRANSACTION tree) and then immediately apply.