Enum db_handling::Error

pub enum Error {

    NoNetworksAvailable,
    DbTransactionError(TransactionError),
    Defaults(Error),
    Other(Error),
    Codec(Error),
    TimeFormat(Format),
    DefinitionsError(Error),
    Bip39MnemonicType(ErrorKind),
    DbError(Error),
    Stub,
    CustomVerifierIsGeneral(VerifierKey),
    DeadVerifier(VerifierKey),
    UnexpectedGenesisHash {
        name: String,
        genesis_hash: H256,
    },
    HistoryEntryNotFound(u32),
    TwoRootKeys {
        seed_name: String,
        encryption: Encryption,
    },
    SeedNameNotMatching {
        address_key: AddressKey,
        expected_seed_name: String,
        real_seed_name: String,
    },
    Qr(String),
    NetworkSpecsKeyForAddressNotFound {
        network_specs_key: NetworkSpecsKey,
        address_key: AddressKey,
    },
    EmptySeed,
    EmptySeedName,
    SecretStringError(SecretStringError),
    KeyCollisionBatch {
        seed_name_existing: String,
        seed_name_new: String,
        cropped_path_existing: String,
        cropped_path_new: String,
        in_this_network: bool,
    },
    KeyCollision {
        seed_name: String,
    },
    DerivationExists {
        multisigner: MultiSigner,
        address_details: Box<AddressDetails>,
        network_specs_key: NetworkSpecsKey,
    },
    InvalidDerivation(String),
    LostPwd,
    DerivationsNotFound,
    Sign,
    NoValidCurrentVerifier(VerifierKey),
    DifferentNamesSameGenesisHash {
        name1: String,
        name2: String,
        genesis_hash: H256,
    },
    DifferentBase58Specs {
        base58_1: u16,
        base58_2: u16,
        genesis_hash: H256,
    },
    GeneralVerifierNotFound,
    TypesNotFound,
    NetworkSpecsNotFound(NetworkSpecsKey),
    AddressNotFound(AddressKey),
    MetaValuesNotFound {
        name: String,
        version: u32,
    },
    ChecksumMismatch,
    DangerStatusNotFound,
    NoKnownSeeds,
    SecretExposedMismatch {
        multisigner: MultiSigner,
        address_details: AddressDetails,
    },
    WrongPassword,
    MissingPasswordInfo(String),
    NoSeedForKeyPair {
        multisigner: MultiSigner,
    },
    NoRootKeyForSeed(String),

}

DB handling error type.

Variants

NoNetworksAvailable

DbTransactionError(TransactionError)

Defaults(Error)

Other(Error)

Codec(Error)

TimeFormat(Format)

DefinitionsError(Error)

Bip39MnemonicType(ErrorKind)

DbError(Error)

Stub

Temporary database entry in TRANSACTION tree of the Vault database under the key STUB, used to store the update data awaiting for the user approval.

Missing Stub when it is expected always indicates the database corruption.

CustomVerifierIsGeneral(VerifierKey)

Network CurrentVerifier is ValidCurrentVerifier::Custom(_), but the custom verifier value coincides with the general verifier.

Associated data is [VerifierKey] corresponding to faulty entry.

DeadVerifier(VerifierKey)

User tried to interact with previously disabled network.

Associated data is the [VerifierKey] of the network.

UnexpectedGenesisHash

Fields

name: String

genesis_hash: H256

Network has no entry for CurrentVerifier under verifier_key in VERIFIERS tree of the database, however, the corresponding genesis hash is encountered in a OrderedNetworkSpecs entry under network_specs_key in SPECSTREE tree of the database. No network specs record can get into database without the verifier entry, and the verifier could not be removed while network specs still remain, so this indicates the database got corrupted.

HistoryEntryNotFound(u32)

History log Entry stored in HISTORY tree of the Vault database under the key Order could not be decoded.

Associated data is the corresponding Order.

TwoRootKeys

Fields

seed_name: String

seed name

encryption: Encryption

encryption algorithm for which two seed keys were found

Database has two seed addresses (i.e. with empty derivation path and no password) for same seed name and [Encryption]

This indicates the database corruption, since the encryption method, seed name and derivation path strictly determine the public key.

SeedNameNotMatching

Fields

address_key: AddressKey

address key for which the export is done

expected_seed_name: String

seed name, from the navigator

real_seed_name: String

seed name, from the AddressDetails

To generate QR code with public address information export, Vault receives both seed name and MultiSigner from the navigation state Navstate. MultiSigner gets transformed into [AddressKey] and corresponds to AddressDetails that are exported. AddressDetails also contain seed_name field, that must coincide with the one received directly from the navigator. This error appears if the seed names are different.

Qr(String)

QR handling error.

NetworkSpecsKeyForAddressNotFound

Fields

network_specs_key: NetworkSpecsKey

address_key: AddressKey

[NetworkSpecsKey] of a network in network_id field of the AddressDetails corresponding to [AddressKey].

This happens when the derivation is created in some other network(s), but not in the given network. This way the AddressKey is in the database, but the address in the network is not.

EmptySeed

Got empty seed phrase.

EmptySeedName

Got empty seed name.

SecretStringError(SecretStringError)

Error in SecretString.

SecretString consists of combined seed phrase and derivation.

Associated error content is SecretStringError.

KeyCollisionBatch

Fields

seed_name_existing: String

seed_name_new: String

cropped_path_existing: String

cropped_path_new: String

in_this_network: bool

Same public key was produced for a different seed phrase and/or derivation path, during database transaction preparation (not yet in the database).

KeyCollision

Fields

seed_name: String

Same public key was produced for a different seed phrase and/or derivation path, as already existing in the database.

Address is generated within a network using seed phrase and derivation path.

Address is defined by public key and [NetworkSpecsKey]. Public key is created from seed phrase and derivation with encryption algorithm supported by the network.

If two networks are using the same encryption algorithm, generating public key in both with same seed phrase and derivation path would result in two identical public keys. This is normal and expected behavior, and this is the reason why [AddressDetails] contain a set of allowed networks in network_id field.

It is, however, possible, that when generating public key for different seed names and/or different derivation paths, resulting public keys accidentally coincide.

This is called here KeyCollision, and results in error.

DerivationExists

Fields

multisigner: MultiSigner

address_details: Box<AddressDetails>

network_specs_key: NetworkSpecsKey

Derivation that user tried to create already exists.

Associated error content is:

MultiSigner of the already existing address

[AddressDetails] for already existing address

[NetworkSpecsKey] of the associated network

InvalidDerivation(String)

Received derivations update payload contains an invalid derivation.

Associated data is the derivation that could not be used as a String.

LostPwd

User was creating the derivation with password, and thus moved into PasswordConfirm modal, however, the password was not found when cutting password from the path.

DerivationsNotFound

Temporary database entry in TRANSACTION tree of the Vault database under the key DRV, used to store the derivation import data.

Missing Derivations when it is expected always indicates the database corruption.

Sign

Temporary database entry in TRANSACTION tree of the Vault database under the key SIGN, used to store the signable transaction data awaiting for the user approval.

Missing Sign when it is expected always indicates the database corruption.

NoValidCurrentVerifier(VerifierKey)

DifferentNamesSameGenesisHash

Fields

name1: String

name2: String

genesis_hash: H256

While searching for all networks with same genesis hash, found that there are networks with same genesis hash, but different names.

DifferentBase58Specs

Fields

base58_1: u16

base58_2: u16

genesis_hash: H256

Network specs entries have same genesis hash, but different base58 prefix

GeneralVerifierNotFound

General verifier Verifier information stored in SETTREE tree of the database under key GENERALVERIFIER.

Missing general verifier always indicates the database corruption.

TypesNotFound

DangerRecord information stored in SETTREE tree of the database under key DANGER.

Missing DangerRecord always indicates the database corruption.

NetworkSpecsNotFound(NetworkSpecsKey)

OrderedNetworkSpecs for a network in SPECSTREE tree of the Vault database, searched by [NetworkSpecsKey].

Associated data is the NetworkSpecsKey used for the search.

AddressNotFound(AddressKey)

AddressDetails for [AddressKey] in ADDRTREE tree of the Vault database.

Associated data is the AddressKey used for search.

MetaValuesNotFound

Fields

name: String

network name

version: u32

network version

Network metadata in METATREE tree of the Vault database, for network name and version combination.

ChecksumMismatch

Database checksum does not match the expected value.

DangerStatusNotFound

DangerRecord information stored in SETTREE tree of the database under key DANGER.

Missing DangerRecord always indicates the database corruption.

NoKnownSeeds

SecretExposedMismatch

Fields

multisigner: MultiSigner

address_details: AddressDetails

Found secret_exposed flag mismatch in the database: address is not marked as potentially exposed when it must have been.

WrongPassword

User has entered a wrong password for a passworded address.

MissingPasswordInfo(String)

NoSeedForKeyPair

Fields

multisigner: MultiSigner

NoRootKeyForSeed(String)