db_handling

Function default_cold_release

pub fn default_cold_release(database: Option<&Db>) -> Result<()>

Generate or restore “cold” database with default values, for release build.

Resulting database should be copied verbatim into Vault files during the build.

The location of the generated database is either optional user-provided path, or default [COLD_DB_NAME_RELEASE] folder.

The cold release database, as generated, contains:

network specs for default networks (Polkadot, Kusama, Westend)

verifier information for default networks, with verifiers set to the general one

two latest metadata versions for default networks

default types information and clean danger status

The trees ADDRTREE, HISTORY, and TRANSACTION are empty.

Note that resulting database history is not initialized and general verifier is not set.

This operation is performed not on Vault device, and is governed by the active side.