Function db_handling::default_hot

pub fn default_hot(path: Option<PathBuf>) -> Result<(), ErrorActive>

Generate or restore “hot” database with default values.

The location of the generated database is default HOT_DB_NAME folder.

The hot database, as generated, contains:

address book entries for default networks (Polkadot, Kusama, Westend)

network specs for default networks

default types information

no metadata entries

no metadata block history entries

All metadata-related entries get in the hot database only through rpc calls.