Function db_handling::helpers::genesis_hash_in_specs

pub fn genesis_hash_in_specs(
    genesis_hash: H256, 
    database: &Db
) -> Result<Option<SpecsInvariants>, ErrorSigner>

Search for network genesis hash in NetworkSpecs entries in SPECSTREE of the Signer database.

Genesis hash is calculated from network VerifierKey input.

If there are more than one network corresponding to the same genesis hash, outputs network specs key for the network with the lowest order.

If there are several entries with same genesis hash, all of them must have identical base58 prefix and network name. Network name is, and base58 prefix could be a part of the network metadata, and therefore must not depend on encryption used.