Function db_handling::helpers::remove_network

pub fn remove_network(
    database: &Db,
    network_specs_key: &NetworkSpecsKey
) -> Result<()>

Remove the network from the database.

Function inputs [NetworkSpecsKey] of the network, gets network genesis hash and proceeds to act on all networks with same genesis hash.

Removing network is mostly an emergency tool and is not expected to be used really often.

Removing a network means:

Remove from SPECSTREE all [OrderedNetworkSpecs] that have genesis hash associated with given NetworkSpecsKey

Remove from METATREE all metadata entries corresponding to the network name, as found in OrderedNetworkSpecs

Remove from ADDRTREE all addresses in the networks being removed

Modify Verifier data if necessary.

Removing the network may result in blocking the network altogether until the Vault is reset. This happens only if the removed network [ValidCurrentVerifier] was set to ValidCurrentVerifier::Custom(Verifier(Some(_))) and is a security measure. This should be used to deal with compromised Custom verifiers.

Compromised general verifier is a major disaster and will require Vault reset in any case.

Removing the network verified by the general verifier does not block the network.

Removing the network verified by ValidCurrentVerifier::Custom(Verifier(None)) does not block the network.

Note that if the network supports multiple encryption algorithms, the removal of network with one of the encryptions will cause the networks with other encryptions be removed as well.