db_handling

Module identities

Creating, collecting, and removing addresses in cold database

In Vault and Vault-related ecosystem the address is a key pair within a certain network. Vault Rust database stores only non-secret data associated with each address generated by user.

Required data to generate address:

seed phrase (secret words set, must be always zeroized after use and is never stored in Rust database)

derivation path (/soft//hard///password); password part, i.e. whatever follows /// is not stored in Rust database and must be zeroized after use

[NetworkSpecsKey]

The combination of seed phrase and derivation path allows to generate key pair using the encryption algorithm supported by the network, for which the address is generated.

Vault stores only the public key, it is used together with encryption algorithm [Encryption] to generate [AddressKey].

Same AddressKey could correspond to more than one network if the addresses were produced for same seed phrase and derivation path, and networks use same encryption algorithm. The networks specs keys of associated networks are stored in network_id field of AddressDetails.

Vault interface shows those addresses as separate entities.

Non-secret data associated with AddressKey is stored in [ADDRTREE] tree of the cold database as SCALE-encoded [AddressDetails].

Structs

AddrInfo

DynamicDerivationTransaction

ExportAddrsV1

ExportAddrsV2

SeedInfo

SignaturesBulkV1

TransactionBulkV1

Enums

DerivationCheck

Proposed derivation status.

ExportAddrs

SignaturesBulk

TransactionBulk

VaultFeatures

Functions

check_derivation_set

Check derivations before offering user to import them.

create_increment_set

Add a set of new derived addresses: N+1, N+2, etc into Vault database.

create_key_set

Creates seed into Vault: add default addresses for derive_for_networks in the Vault

cut_path

Separate derivation path with password into non-secret part and password.

derivation_check

Preliminary checks if the user-proposed derivation could be used to create the address.

derive_single_key

Helper function to get public key from seed phrase and derivation path

dynamic_derivations_response

Prepare DynamicDerivationsAddressResponse for the given request

export_key_set_addrs

Export info about keys and their addresses known to Vault

export_secret_key

Prepare secret key export screen struct [MKeyDetails].

find_address_details_for_multisigner

Return address details that corresponds to a public key. The root public keys are prioritized. Then network hashes passed as last param allows to choose one address in case multiple addresses available in different networks

generate_random_phrase

Generate random phrase with given number of words.

generate_test_identities

Generate test Alice addresses in test cold database.

get_addresses_by_seed_name

Get all existing addresses for a given seed name from the database.

get_all_addresses

Get all existing addresses from the database.

get_multisigner_by_address

import_all_addrs

inject_derivations_has_pwd

is_passworded

Check derivation format and determine if it has a password.

process_dynamic_derivations_v1

remove_key

Remove address from the Vault database.

remove_keys_set

Remove a set of addresses within a single network from the Vault database.

remove_seed

Remove all addresses associated with given seed name from the Vault database.

try_create_address

Create a new address in the Vault database.

try_create_seed

Accept seed into Vault: add default addresses for a seed in the Vault database.

validate_key_password