db_handling::identities

Function check_derivation_set

pub fn check_derivation_set(derivations: &[String]) -> Result<()>

Check derivations before offering user to import them.

Vault allows bulk import only for valid derivations, with or without password. For passworded derivations no password is ever confirmed, as the user presumably has the derivations already recorded somewhere.

This function is used in transaction_parsing::derivations to check the incoming derivations once they are received in derivations qr code payload.

Note that once the derivations are checked, they are moved around as “checked derivations” and not checked anymore until they are used in import_derivations. Only checked derivations are written into the database temporary storage in TRANSACTION tree, same data is recovered as ensured by the checksum matching.