db_handling::identities

Function create_increment_set

pub fn create_increment_set(
    database: &Db,
    increment: u32,
    multisigner: &MultiSigner,
    network_specs_key: &NetworkSpecsKey,
    seed_phrase: &str,
) -> Result<()>

Add a set of new derived addresses: N+1, N+2, etc into Vault database.

The number of the added addresses is set by increment input.

Input [MultiSigner] and [NetworkSpecsKey] determine the address, and the associated [AddressDetails] are found in the database. The derivations are based on the path from the AddressDetails.

If the database has addressed in given network with path that could be interpreted as <path>//<number>, the first derivation of the increment set is <path>//<number>+1. If no such addresses are existing in the database, the first derivation of the increment set is <path>//0.

Further derivations of the increment set have number part increased by 1 until all requested derivations are generated.

Note: this is intended only for addresses not protected by the password (i.e. having derivation path without ///<password> part). If the increment set is created on a passworded address, only public derivation path part will be used for increment set generation.

This function inputs secret seed phrase as &str. It is passed as &str into create_address and used there.

Example

User creates increment set of 3 on derivation //user. Resulting derivations are:

//user

//user//0

//user//1

//user//2

User creates increment set of 2 on //user again, resulting derivations are:

//user

//user//0

//user//1

//user//2

//user//3

//user//4

If the increment set of 2 is created on //user//0, resulting derivations would be:

//user

//user//0

//user//0//0

//user//0//1

//user//1

//user//2

//user//3

//user//4