db_handling::identities

Function export_secret_key

pub fn export_secret_key(
    database: &Db,
    public_key: &str,
    expected_seed_name: &str,
    network_specs_key_hex: &str,
    seed_phrase: &str,
    key_password: Option<String>,
) -> Result<MKeyDetails>

Prepare secret key export screen struct [MKeyDetails].

For QR code the secret key information is put in format secret:0x{}:{network genesis hash} transformed into bytes, to be compatible with polkadot-js interface.

There is no direct mentioning of [Encryption] in the QR code. It is that user knows the key in which network and with what encryption they are exporting.

The QR code here contains sensitive information, and is made with special coloration, so that the difference with safe QR codes is immediately visible on screen.