db_handling::identities

Function remove_keys_set

pub fn remove_keys_set(
    database: &Db,
    multiselect: &[MultiSigner],
    network_specs_key: &NetworkSpecsKey,
) -> Result<()>

Remove a set of addresses within a single network from the Vault database.

Address is determined by [MultiSigner] and [NetworkSpecsKey] of the network. Function inputs network NetworkSpecsKey and a set of MultiSigner values for which the address within a network must be deleted.

NetworkSpecsKey of the network is removed from network_id set in [AddressDetails] associated with each of the provided MultiSigner values. If no networks associated with [AddressKey] remain, i.e. network_id set becomes empty, whole associated entry is removed.