db_handling

Module interface_signer

Utils to communicate with the Vault frontend

Structs

SeedDraft

Zeroizeable seed phrase draft.

Constants

BIP_CAP

Maximum word count in bip39 standard.

WORD_LENGTH

Maximum word length in bip39 standard.

Functions

addresses_set_seed_name_network

Get address-associated public data for all addresses from the Vault database with given seed name and network [NetworkSpecsKey].

backup_prep

Prepare seed backup screen struct [MBackup] for given seed name.

derive_prep

Prepare key derivation screen struct [MDeriveKey].

dynamic_path_check

Return [NavDerivationCheck] with allowed action details for new key derivation.

export_key

Prepare export key screen struct [MKeyDetails].

first_network

Sort database networks by the order and get the network specs for the first network on the list.

get_all_seed_names_with_identicons

Return a Vec with all seed names with seed key identicons if seed key is available.

guess

Get possible options of English bip39 words that start with user-entered word part.

history_hex_checksum

Get database history tree checksum to be displayed in log screen.

keys_by_seed_name

metadata_details

Return [MManageMetadata] with metadata details for network with given [NetworkSpecsKey] and given version.

network_details_by_key

Return [MNetworkDetails] with network specs and metadata set information for network with given [NetworkSpecsKey].

print_all_identities

Return a Vec with address-associated public data for all addresses from the Vault database.

print_new_seed

Generate new random seed phrase, make identicon for sr25519 public key, and send to Vault screen.

purge_transactions

Clear transaction tree of the database.

show_all_networks

Make Vec with network information for all networks in the Vault database, without any selection.

show_all_networks_with_flag

Return Vec with network information for all networks in the Vault database, with bool indicator which one is currently selected.

show_types_status

Make types status display.