Function db_handling::interface_signer::get_all_seed_names_with_identicons

pub fn get_all_seed_names_with_identicons(
    database_name: &str, 
    names_phone_knows: &[String]
) -> Result<Vec<SeedNameCard>, ErrorSigner>

Return a Vec with all seed names with seed key identicons if seed key is available.

Function processes all seeds known to Signer KMS (which are input as &[String]), including seeds without any corresponding addresses currently known to Signer (orphans).

If the same seed has more than one seed key in the database, i.e. it has been used to create seed keys with more than one Encryption algorithm, only one identicon is selected, in order of preference: Sr25519, Ed25519, Ecdsa.