db_handling

Module ledger_ed25519

Ledger Ed25519 key derivation using Cardano-style BIP32-Ed25519

This implements the same key derivation used by Ledger hardware wallets for Polkadot and other Substrate-based chains.

Key derivation path: m/44’/slip0044’/account’/change’/address’ For Polkadot: m/44’/354’/0’/0’/0’

Note: Ledger uses Cardano-style BIP32-Ed25519, NOT SLIP-10. The difference is in the master key generation which iterates until a valid key is found.

Structs

LedgerKeyPair

Ledger key pair with extended key for signing

Constants

SLIP0044_KUSAMA

SLIP0044_POLKADOT

SLIP-0044 coin types

Functions

derive_kusama_ledger_key

Derive Kusama Ledger key (m/44’/434’/account’/0’/0’)

derive_ledger_key

Full Ledger key derivation from mnemonic Path: m/44’/slip0044’/account’/change’/address’

derive_polkadot_ledger_key

Derive Polkadot Ledger key (m/44’/354’/account’/0’/0’)