db_handling

Module nostr

Nostr key derivation using NIP-06 and event signing

This implements key derivation for Nostr using:

NIP-06: Basic key derivation from mnemonic (m/44’/1237’/account’/0/0)

BIP-340: Schnorr signatures for event signing

Nostr uses x-only public keys (32 bytes) and Schnorr signatures.

Structs

NostrKeyPair

Nostr key pair with x-only public key

SignedEvent

Signed Nostr event (NIP-01)

UnsignedEvent

Unsigned Nostr event (NIP-01)

Constants

COIN_TYPE_NOSTR

SLIP-0044 coin type for Nostr (NIP-06)

Functions

derive_nostr_key

Derive Nostr key from seed phrase (NIP-06)

event_kind_name

Get human-readable event kind name

get_nostr_address

Retrieve a Nostr npub for a given public key hex

parse_nostr_path

Parse Nostr derivation path

requires_confirmation

Event kinds that require explicit user confirmation

store_nostr_address

Store a Nostr npub for a given public key hex