Crate db_handling

This crate deals with standard database-related procedures used in Vault and Vault-supporting ecosystem.

This crate:

contains helpers to operate the databases, used throughout Vault system

generates cold (used in air-gapped Vault) and hot (used in generate_message client) databases with default settings

deals with taking data out of the database and putting the data in the database

contains Vault interface interactions, allowing exports of data to the interface

deals with address generation for Vault

Features

Feature "active" corresponds to all Vault-related things happening without air-gap, including the generation of the database for Vault during the build.

Modules

cold_default

Making and restoring cold database with default content

db_transactions

Atomic transactions in cold and hot databases

helpers

Common helper functions for database operations

identities

Creating, collecting, and removing addresses in cold database

interface_signer

Utils to communicate with the Vault frontend

manage_history

Displaying and updating history log

Enums

Error

DB handling error type.

Functions

default_cold_release

Generate or restore “cold” database with default values, for release build.

default_hot

Generate or restore “hot” database with default values.

Type Definitions

Result

DB handling result.