Crate db_handling

This crate deals with standard database-related procedures used in Signer and Signer-supporting ecosystem.

This crate:

contains helpers to operate the databases, used throughout Signer system

generates cold (used in air-gapped Signer) and hot (used in generate_message client) databases with default settings

deals with taking data out of the database and putting the data in the database

contains Signer interface interactions, allowing exports of data to the interface

deals with address generation for Signer

Features

Feature "signer" corresponds to everything happening in Signer air-gapped device.

Feature "active" corresponds to all Signer-related things happening without air-gap, including the generation of the database for Signer during the build.

Feature "test" includes both "signer" and "active" features, along with some testing, and is the default one.

Modules

cold_default

Making and restoring cold database with default content

db_transactions

Atomic transactions in cold and hot databases

helpers

Common helper functions for database operations

identities

Creating, collecting, and removing addresses in cold database

interface_signer

Utils to communicate with the Signer frontend

manage_history

Displaying and updating history log

Functions

default_cold_release

Generate or restore “cold” database with default values, for release build.

default_hot

Generate or restore “hot” database with default values.