db_handling

Module db_transactions

Atomic transactions in cold and hot databases

Additions and removals of entries in cold and hot database occur through atomic transactions. Each tree gets updated with its own [Batch], updates occur within a single transaction.

For transactions scanned into Vault, currently a temporary database entry is made to store transaction details while they are displayed to user.

Structs

TrDbCold

Cold database transaction data containing [Batch] elements that will be applied to each Tree.

TrDbColdSign

Temporary storage for signable transaction and associated data.

TrDbColdSignOne

TrDbColdStub

Draft for cold database atomic transaction, constructed for Vault update transaction (add_specs, load_metadata, load_types).

TrDbHot

Hot database transaction data containing [Batch] elements that will be applied to each Tree.

Enums

SignContent

Signable transaction content