Module db_handling::helpers

Common helper functions for database operations

Structs

SpecsInvariants

Specs invariants that are expected to stay unchanged for the network over time and can not be different for same genesis hash and different encryption algorithms.

Functions

assert_db_version

Check database schema version. Prevents app stash when user upgrades the app without re-install.

genesis_hash_in_specs

Search for network genesis hash in OrderedNetworkSpecs entries in SPECSTREE of the Vault database.

get_address_details

Get AddressDetails from the Vault database, using AddressKey.

get_all_networks

Get OrderedNetworkSpecs for all networks from the cold database.

get_danger_status

Get the danger status from the Vault database.

get_general_verifier

Get general verifier Verifier from the Vault database.

get_meta_values_by_name_version

Get MetaValues, corresponding to given network name and version, from the database.

get_network_specs

Get network specs OrderedNetworkSpecs from the Vault database.

get_types

Get types information as Vec<TypeEntry> from the database.

get_valid_current_verifier

Get ValidCurrentVerifier from the Vault database for a network, using VerifierKey.

make_batch_clear_tree

Assemble a [Batch] that removes all elements from a tree.

open_tree

Open a tree in the database.

prep_types

Get types information as ContentLoadTypes from the database.

remove_metadata

Remove the network metadata entry from the database.

remove_network

Remove the network from the database.

remove_types_info

User-initiated removal of the types information from the Vault database.

transfer_metadata_to_cold

Transfer metadata from the hot database into the cold one.

try_get_address_details

Try to get AddressDetails from the Vault database, using AddressKey.

try_get_meta_values_by_name_version

Try to get MetaValues, corresponding to given network name and version from the database.

try_get_network_specs

Try to get network specs OrderedNetworkSpecs from the Vault database.

try_get_types

Try to get types information from the database.

try_get_valid_current_verifier

Try to get ValidCurrentVerifier from the Vault database for a network, using VerifierKey.

validate_mnemonic