Module db_handling::identities

Creating, collecting, and removing addresses in cold database

In Signer and Signer-related ecosystem the address is a key pair within a certain network. Signer Rust database stores only non-secret data associated with each address generated by user.

Required data to generate address:

seed phrase (secret words set, must be always zeroized after use and is never stored in Rust database)

derivation path (/soft//hard///password); password part, i.e. whatever follows /// is not stored in Rust database and must be zeroized after use

NetworkSpecsKey

The combination of seed phrase and derivation path allows to generate key pair using the encryption algorithm supported by the network, for which the address is generated.

Signer stores only the public key, it is used together with encryption algorithm Encryption to generate AddressKey.

Same AddressKey could correspond to more than one network if the addresses were produced for same seed phrase and derivation path, and networks use same encryption algorithm. The networks specs keys of associated networks are stored in network_id field of AddressDetails.

Signer interface shows those addresses as separate entities.

Non-secret data associated with AddressKey is stored in ADDRTREE tree of the cold database as SCALE-encoded AddressDetails.

Enums

DerivationCheck

Proposed derivation status.

Functions

check_derivation_set

Check derivations before offering user to import them.

create_increment_set

Add a set of new derived addresses: N+1, N+2, etc into Signer database.

cut_path

Separate derivation path with password into non-secret part and password.

derivation_check

Preliminarily checks if the user-proposed derivation could be used to create the address.

generate_random_phrase

Generate random phrase with given number of words.

generate_test_identities

Generate test Alice addresses in test cold database.

get_addresses_by_seed_name

Get all existing addresses for a given seed name from the database.

import_derivations

Create a set of addresses using imported derivations set for user-selected seed.

prepare_derivations_import

Prepare derivations import content ContentDerivations.

remove_key

Remove address from the Signer database.

remove_keys_set

Remove a set of addresses within a single network from the Signer database.

remove_seed

Remove all addresses associated with given seed name from the Signer database.

try_create_address

Create a new address in the Signer database.

try_create_seed

Accept seed into Signer: add default addresses for a seed in the Signer database.