Function db_handling::identities::derivation_check

pub fn derivation_check(
    seed_name: &str, 
    path: &str, 
    network_specs_key: &NetworkSpecsKey, 
    database_name: &str
) -> Result<DerivationCheck, ErrorSigner>

Preliminarily checks if the user-proposed derivation could be used to create the address.

This function does not use secret phrase, and only searches the AddressDetails from the database for already existing addresses with same derivation path, in case there is no password in the proposed derivation.

If the proposed derivation is passworded, no checking could be done without using the seed phrase, and UI allows to proceed to try creating new address. If address already exists, an error would appear on the next stage.

No checking for collisions is done here (i.e. no checking if the same AddressKey will be produced for <secret phrase 1> + <derivation path 1> as for already known <secret phrase 2> + <derivation path 2>), because seed phrase is not known. If collision occurs, an error would appear on the next stage.

Note that the only errors in this function could appear from database malfunction when the database is being searched for the exact match, in case of password-free valid derivation. Bad format of the derivation is not an error, UI just does not allow to proceed.