Function db_handling::identities::prepare_derivations_import

pub fn prepare_derivations_import(
    encryption: &Encryption, 
    genesis_hash: H256, 
    content: &str
) -> Result<ContentDerivations, ErrorActive>

Prepare derivations import content ContentDerivations.

Function is used on the active side. It inputs information about network in which the derivations would be created (genesis hash and Encryption), and the plaintext user derivations set.

Derivations allowed in the import set must be valid, with or without password. The derivations source file must have derivations as a list with each new derivation on the new line.

The source plaintext is cut in lines and each line is processes using regex, all invalid derivations are ignored. Function prints found valid password-free derivations as is and passworded with only public part visible (e.g. //alice///<password>).