Function db_handling::identities::try_create_seed

pub fn try_create_seed(
    database: &Db,
    seed_name: &str,
    seed_phrase: &str,
    make_seed_keys: bool
) -> Result<()>

Accept seed into Vault: add default addresses for a seed in the Vault database.

Each network has a default derivation path, recorded in [NetworkSpecs] field path_id, normally path_id is //<network_name>.

This function adds addresses with default derivation paths into the Vault database. If the make_seed_keys flag is set to true, seed keys are also added. History log is accordingly updated.

This function inputs secret seed phrase as &str. It is passed as &str into populate_addresses and used there.