Function db_handling::identities::try_create_seed

pub fn try_create_seed(
    seed_name: &str, 
    seed_phrase: &str, 
    make_seed_keys: bool, 
    database_name: &str
) -> Result<(), ErrorSigner>

Accept seed into Signer: add default addresses for a seed in the Signer database.

Each network has a default derivation path, recorded in NetworkSpecs field path_id, normally path_id is //<network_name>.

This function adds addresses with default derivation paths into the Signer database. If the make_seed_keys flag is set to true, seed keys are also added. History log is accordingly updated.

This function inputs secret seed phrase as &str. It is passed as &str into populate_addresses and used there.