Module db_handling::manage_history

Displaying and updating history log

The Vault keeps a history log of all events that change the database and affect security. It is stored in HISTORY tree of the cold database.

Each history log Entry contains Event set and a timestamp. Database key for Entry value is Order, SCALE-encoded number of the entry.

In addition to keeping the log, Vault also displays HISTORY tree checksum for user to possibly keep the track of.

Functions

clear_history

Clear Vault history and make a log Entry that history was cleared.

device_was_online

Process the fact that the Vault device was online.

enter_events

Enter Event set into the database as a single database transaction.

events_to_batch

Timestamp Event set and make with it a new [Batch], that could be applied to the HISTORY tree.

get_history

Get history log contents from the database.

get_history_entry_by_order

Get from the database a history log Entry by u32 order identifier received from the frontend.

history_entry_system

Enter system-generated Event into the database.

history_entry_user

Enter user-generated Event into the database.

history_total_pages

Print total number of pages, for maximum HISTORY_PAGE_SIZE number of entries per page.

reset_danger_status_to_safe

Acknowledge that the Vault device was online and reset the DangerRecord back to safe.

seed_name_was_shown

Record in history log that certain seed was shown on Vault screen, presumably for backup.